+++ sapp/context_manager.py
@@ -1,7 +1,7 @@
 from collections.abc import Iterable
 
 
-class ContextManager(object):
+class ContextManager:
     def __init__(self, application, values=[]):
         self.application = application
         self.values = values
@@ -20,3 +20,26 @@ class ContextManager(object):
 
     def __exit__(self, exc_type, exc_value, traceback):
         self.application._exit_context(exc_type, exc_value, traceback)
+
+
+class LazyContextManager:
+    def __init__(self, application):
+        self.application = application
+        self._context = None
+
+    @property
+    def context(self):
+        if not self._context:
+            self._context = self.application._enter_context()
+
+        return self._context
+
+    def get(self, parameter):
+        return getattr(self.context, parameter)
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        if self._context:
+            self.application._exit_context(exc_type, exc_value, traceback)